Make testing simpler
